Responsive status area layout (#69)
* Add bottom bar icons

* Add icons to layout

* Remove programatic icon setting

* Convert status.xml to use Constraint Layouts

* Convert status.xml to use Constraint Layouts
https://developer.android.com/training/constraint-layout
This should allow the area to be "responsive" and adjust itself based on
the space available

* The "Engine status" and "Opening" lines are now on separate horizontal
lines.

* Change the colon color to blue, so it doesn't look like it is owned by
white. (Do we even need the colon to be there?)

* Set maximum font size for disk count

This sets the maximum font size that the auto-scaling disk count digits
can be.

80sp is an arbitrary value that was "not too big"

* Prevent line-wrapping on the disc counts.

Set maxlines to 1 to prevent line-wrapping on the disc counts when the
auto-scaled font fits, but the 'to move' indicator then causes the line
to want to wrap.